Rework waves system + windows + music
Shadows spawn permanently. There can be only one at a time at first.
When you open a curtain a wave starts and they spawn more and more often.
After a wave they return to normal but there can more at once and next wave requires more to be killed.

Windows are a class now and put in a node so lighthouse.gd can parse this node and record all windows in the level without manually adding them

Music is a bit higher for better feeling on speaker, still possible to change it in pause